Use a native WebView on OS X
There's a Qt bug (https://bugreports.qt-project.org/browse/QTBUG-22269) that
makes QtWebkit scrolling on OS X basically unusable for us. A couple of workarounds
that I tried didn't work, so this is the bazooka approach and replaces QtWebKit on OSX
with a  native Cocoa webview.

This still doesn't support the bridge, but that is coming. It also doesn't support
cmd-a/c/v/x right now, that is an open bug