Allow image pasting in the desktop app
This is a pile of hacks held together by dental floss, but it works.
Since Safari doesn't support the HTML5 apis we need, we grab an image
from the clipboard in Cocoa, convert it to PNG and Base64 encode it,
then pass it into JS where we call the upload code.

We also replace the sendAsBinary implementation with a polyfill
to avoid a crash deep in JSC.