Swizzle system cookies during websocket handshake. :tada: :voodoo: :dolls:
We are forced to replace the system cookies for zulip with our own cookies,
as when websockets handshake webkit does not use custom supplied cookies.
This means that in order for our websocket handshake to succeed, we temporarily
overwrite the system cookies for this zulip domain with our own cookies for this domain.

We replace them when we are done.

https://bugs.webkit.org/show_bug.cgi?id=124580