Revamp the OS X signing and build process
This properly signs with mac v2 certificates so that the desktop
app appears verified on OS X 10.9+. We are forced to sign on a
10.9+ machine, but we want to build on 10.6 in order to run on
the widest types of OSes possible. So, we scp the Zulip.app bundle
to a 10.9+ machine during the build process, and then scp it back
after signing.

We also now copy Info.plist files for the frameworks we bundle with
our app, a new requirement by Apple for the v2 signing to be successful